Tests/LibWeb/Text/expected/DOM/Element-from-point.txt

Negative coordinates return null: true
Coordinates outside the viewport return null: true
<HTML id="root-element" >
<DIV id="large-box" >
<DIV id="small-box" >
